Add error_correction mode
This allows pysolarmanv5 to detect issues such as the concatenation of
two v5 frames by comparing the v5 frame length to the payload length
defined in the payload_len field.

If error_correction mode is enabled, it will override the v5 frame_len
value with that taken from the payload_len field.

error_correction mode is disabled by default.