Address linter findings, part 2
Disabled `invalid-overridden-method` rule in `pysolarmanv5_async.py`.
`PySolarmanV5Async` overrides sync methods from `PySolarmanV5` that
is used as the base class. The override of a sync method with
an async method in this case is not a problem in practice.

Note: rule disabling could be avoided by changing the class
hierarchy. One way would be to introduce a common base class for
async and sync classes. The new class would capture attributes
and common (class internal) sync methods.